ses: fallback to read SAS address from VPD
The mpt3sas driver does not create a parent end device for PCIe types
where the SAS address is stored, causing the ses driver to not add
PCIe device types connected to a tri-mode HBA. To address this, the
fallback mechanism reads the SAS address from the VPD 0x83 page. This
change is inspired by commit 9927c68.